+++ src/Components/src/Page/FileExplorer/Types.fs
@@ -1,7 +1,6 @@
 module Swate.Components.Page.FileExplorer.Types
 
 open System
-open Feliz
 
 [<RequireQualifiedAccess>]
 type FileItemIconTone =
@@ -241,85 +240,6 @@ module FileTree =
         Selectable = false
         Path = None
     }
-
-type ContextMenuItem = {
-    Label: string
-    Icon: string
-    OnClick: unit -> unit
-    Disabled: bool option
-    ClassName: string option
-    IsDivider: bool option
-} with
-
-    member this.ToPrimitiveContextMenuItem() =
-        if defaultArg this.IsDivider false then
-            Swate.Components.Primitive.ContextMenu.Types.ContextMenuItem(isDivider = true)
-        else
-            let isDisabled = defaultArg this.Disabled false
-
-            let className =
-                [
-                    this.ClassName
-
-                    if isDisabled then
-                        Some "swt:opacity-50"
-                ]
-                |> List.choose id
-                |> String.concat " "
-
-            Swate.Components.Primitive.ContextMenu.Types.ContextMenuItem(
-                text = Html.span [ prop.className className; prop.text this.Label ],
-                icon =
-                    Html.i [
-                        prop.className [
-                            "swt:iconify " + this.Icon
-
-                            if not (System.String.IsNullOrWhiteSpace className) then
-                                className
-                        ]
-                    ],
-                onClick =
-                    (fun _ ->
-                        if not isDisabled then
-                            this.OnClick()
-                    )
-            )
-
-[<RequireQualifiedAccess>]
-module ContextMenuItem =
-
-    let create (label: string) (icon: string) (onClick: unit -> unit) : ContextMenuItem = {
-        Label = label
-        Icon = icon
-        OnClick = onClick
-        Disabled = None
-        ClassName = None
-        IsDivider = None
-    }
-
-    let styled (label: string) (icon: string) (className: string) (onClick: unit -> unit) : ContextMenuItem = {
-        (create label icon onClick) with
-            ClassName = Some className
-    }
-
-    let disabled (label: string) (icon: string) : ContextMenuItem = {
-        (create label icon ignore) with
-            Disabled = Some true
-    }
-
-    let divider: ContextMenuItem = {
-        (disabled "" "") with
-            IsDivider = Some true
-    }
-
-    let forItem label icon onClick item =
-        create label icon (fun () -> onClick item)
-
-    let whenItem predicate label icon onClick item =
-        if predicate item then
-            [ forItem label icon onClick item ]
-        else
-            []
 
 module FileExplorerLogic =