+++ src/Components/src/Primitive/ContextMenu/Types.fs
@@ -15,6 +15,10 @@ type ContextMenuItem
                 label: string
             |},
         ?isDivider: bool,
+        ?disabled: bool,
+        ?className: string,
+        ?label: string,
+        ?iconClass: string,
         ?onClick:
             {|
                 buttonEvent: Browser.Types.MouseEvent
@@ -26,4 +30,16 @@ type ContextMenuItem
     member val icon = icon with get, set
     member val kbdbutton = kbdbutton with get, set
     member val isDivider: bool = defaultArg isDivider false with get, set
+    member val disabled: bool = defaultArg disabled false with get, set
+    member val className = className with get, set
+    member val label = label with get, set
+    member val iconClass = iconClass with get, set
     member val onClick = onClick with get, set
+    member this.Label = defaultArg this.label ""
+    member this.Icon = defaultArg this.iconClass ""
+    member this.Disabled = if this.disabled then Some true else None
+    member this.ClassName = this.className
+    member this.IsDivider = if this.isDivider then Some true else None
+
+    member this.OnClick() =
+        this.onClick |> Option.iter (fun handler -> handler (unbox null))